requirements.txt
alembic==1.8.1
anyio==3.6.1
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bcrypt==4.0.0
bleach==3.2.1
cffi==1.14.3
click==8.1.3
colorama==0.4.4
decorator==4.4.2
defusedxml==0.6.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
entrypoints==0.3
fastapi==0.82.0
greenlet==1.1.3
h11==0.13.0
idna==3.3
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
Mako==1.2.3
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.4
packaging==20.4
pandocfilters==1.4.3
parso==0.7.1
passlib==1.7.4
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.8
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.20
pydantic==1.10.2
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-jose==3.3.0
python-multipart==0.0.5
pywin32==300
pywinpty==0.5.7
pyzmq==20.0.0
qtconsole==4.7.7
QtPy==1.9.0
rsa==4.9
Send2Trash==1.5.0
six==1.15.0
sniffio==1.3.0
SQLAlchemy==1.4.40
starlette==0.19.1
terminado==0.9.1
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
typing_extensions==4.3.0
uvicorn==0.18.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1